Move this to its own file (it's still a nested class though)
